Move Profile serializer and views to api app
