+++ sharelockers/sharelockers/api/urls.py
@@ -1,11 +1,11 @@
 from django.conf.urls import include, url
 from rest_framework import routers
 from api import views
-import profiles.views
+# import profiles.views
 
 router = routers.DefaultRouter()
 router.register(r'lockers', views.LockerViewSet, base_name="lockers")
-router.register(r'users', profiles.views.ProfileViewSet, base_name="profiles")
+router.register(r'users', views.ProfileViewSet, base_name="profiles")
 
 urlpatterns = [
     url(r'^api/', include(router.urls)),
