+++ sharelockers/sharelockers/profiles/models.py
@@ -1,6 +1,5 @@
 from django.db import models
 from django.contrib.auth.models import User
-from rest_framework import serializers
 
 
 class Profile(models.Model):
@@ -9,8 +8,5 @@ class Profile(models.Model):
     description = models.TextField(max_length=255, null=True)
     alias = models.CharField(max_length=255)
 
+    d
 
-class ProfileSerializer(serializers.ModelSerializer): #FIXME: add Hyperlinked
-    class Meta:
-        model = Profile
-        fields = ('user', 'rating', 'description', 'alias')
