+++ sharelockers/sharelockers/profiles/views.py
@@ -3,9 +3,7 @@ from django.contrib.auth import authenticate, login
 from django.shortcuts import redirect
 from django.contrib import messages
 from profiles.models import Profile
-from profiles.models import ProfileSerializer
 from profiles.forms import UserForm, ProfileForm
-from rest_framework import viewsets
 
 
 def user_register(request):
@@ -34,9 +32,4 @@ def user_register(request):
     return render(request, "profiles/register.html", {'user_form': user_form,
                                                       'profile_form': profile_form,
                                                       })
-class ProfileViewSet(viewsets.ModelViewSet):
-    """
-    API endpoint that allows user profiles to be viewed or edited.
-    """
-    queryset = Profile.objects.all()
-    serializer_class = ProfileSerializer
+
